// src/deprecated/shared/Components/Enums/WidgetEnum.ts
export enum WidgetEnum {
  CAROUSEL = 'CAROUSEL',
  HERO_SECTION = 'HERO_SECTION',
  EXAM_CATEGORIES = 'EXAM_CATEGORIES',
  VIDYAPEETH = 'VIDYAPEETH',
  FEATURE_LIST = 'FEATURE_LIST',
  STATS = 'STATS',
  RESULTS = 'RESULTS',
  TESTIMONIALS = 'TESTIMONIALS',
  STUDY_RESOURCE = 'STUDY_RESOURCE',
  YOUTUBE_STATS = 'YOUTUBE_STATS',
}